requirements.txt
backtesting
pandas
numpy
requests
python-binance
matplotlib
keyboard
ta
mplfinance
coinbase-advanced-py
cryptography==42.0.5
PyJWT==2.8.0